decided to move FakeVehicleGenerator into fake_api_client and hence deleted its former module
