+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/data_sources/fake_api_client/FakeVehicleGenerator.kt
@@ -1,4 +1,4 @@
-package com.igor_shaula.fake_data_generator
+package com.igor_shaula.api_polling.data_layer.data_sources.fake_api_client
 
 // TODO this class should produce data similar to what is coming over network - in String/JSON type
